+++ packages/subgraph/src/helpers/updaters.ts
@@ -6,7 +6,7 @@ import {
   Value
 } from "@graphprotocol/graph-ts";
 
-import { CollateralCommitted } from "../../generated/CollateralManager/CollateralManager";
+
 import {
   Bid,
   BidCollateral,
@@ -611,15 +611,6 @@ function setEntityDurationAverage(
   );
 }
 
-export function updateCollateral(
-  collateral: BidCollateral,
-  event: ethereum.Event
-): void {
-  const evt = changetype<CollateralCommitted>(event);
-  collateral.amount = evt.params._amount;
-  collateral.tokenId = evt.params._tokenId;
-  collateral.collateralAddress = evt.params._collateralAddress;
-}
 
 export function replaceLender(bid: Bid, newLender: Lender): void {
   const oldLenderCount = LoanStatusCount.load(`lender-${bid.lender!}`);
