+++ packages/subgraph/src/subgraph.handlebars
@@ -1,3 +1,13 @@
+{{#*inline "global_required_abis"}}
+        - name: TellerV2
+          file: ./abis/{{network}}_TellerV2.json
+        - name: TellerV0Storage
+          file: ../../packages/contracts/generated/artifacts/contracts/TellerV0Storage.sol/TellerV0Storage.json
+        - name: ERC165
+          file: ../../packages/contracts/generated/artifacts/@openzeppelin/contracts/utils/introspection/ERC165.sol/ERC165.json
+        - name: IERC20Metadata
+          file: ../../packages/contracts/generated/artifacts/@openzeppelin/contracts/token/ERC20/extensions/IERC20Metadata.sol/IERC20Metadata.json
+{{/inline}}
 specVersion: 0.0.2
 schema:
   file: ./src/schema.graphql
@@ -251,12 +261,9 @@ dataSources:
         - RewardAllocation
         - ClaimableReward
       abis:
+{{> global_required_abis}}
         - name: MarketLiquidityRewards
           file: ./abis/{{network}}_MarketLiquidityRewards.json
-        - name: ERC165
-          file: ../../packages/contracts/generated/artifacts/@openzeppelin/contracts/utils/introspection/ERC165.sol/ERC165.json
-        - name: IERC20Metadata
-          file: ../../packages/contracts/generated/artifacts/@openzeppelin/contracts/token/ERC20/extensions/IERC20Metadata.sol/IERC20Metadata.json
       eventHandlers:
         - event: CreatedAllocation(uint256,address,uint256)
           handler: handleCreatedAllocation
